+++ Source/Classes/PINRemoteImageCaching.h
@@ -42,6 +42,10 @@ typedef void (^PINRemoteImageCachingObjectBlock)(id<PINRemoteImageCaching> cache
 @optional
 
 - (void)removeObjectForKeyFromMemory:(NSString *)key;
+- (void)setObjectInMemory:(id)object forKey:(NSString *)key withCost:(NSUInteger)cost withAgeLimit:(NSTimeInterval)ageLimit;
+- (void)setObjectOnDisk:(id)object forKey:(NSString *)key withAgeLimit:(NSTimeInterval)ageLimit;
+- (BOOL)memoryCacheIsTTLCache;
+- (BOOL)diskCacheIsTTLCache;
 
 @end